+++ cli/azd/pkg/infra/provisioning/manager.go
@@ -58,8 +58,19 @@ func (m *Manager) Deploy(ctx context.Context) (*DeployResult, error) {
 		return nil, fmt.Errorf("error deploying infrastructure: %w", err)
 	}
 
+	skippedDueToDeploymentState := deployResult.SkippedReason == DeploymentStateSkipped
+
+	if skippedDueToDeploymentState {
+		m.console.StopSpinner(ctx, "Didn't find new changes.", input.StepSkipped)
+		m.console.ShowSpinner(ctx, "Restore Azure Deployment State.", input.Step)
+	}
+
 	if err := UpdateEnvironment(m.env, deployResult.Deployment.Outputs); err != nil {
 		return nil, fmt.Errorf("updating environment with deployment outputs: %w", err)
+	}
+
+	if skippedDueToDeploymentState {
+		m.console.StopSpinner(ctx, "Restore Azure Deployment State.", input.GetStepResultFormat(err))
 	}
 
 	// make sure any spinner is stopped